+++ src/test-acceptance/java/jscover/server/HtmlUnitServerTest.java
@@ -371,7 +371,7 @@ public class HtmlUnitServerTest {
             "-ws",
             "--document-root=src/test-acceptance/resources",
             "--port=9001",
-            "--no-instrument=/example/lib",
+            "--no-instrument=example/lib",
             "--report-dir=" + reportDir
     };